requirements.txt
flask
requests
flask-sqlalchemy
flask-wtf
flask-login
flask-dance[sqla]
blinker
flask-bcrypt
flask-mail
pillow
